+++ modules/Users/Exceptions.php
@@ -384,14 +384,14 @@ if ( UserStaffID()
 
 		echo '<br />';
 
-		$error[] = sprintf(
+		$warning[] = sprintf(
 			_( '%s is assigned to the profile %s.' ),
 			$staff_RET[1]['FULL_NAME'],
-			$profile_title[1]['TITLE']
+			'"' . _( $profile_title[1]['TITLE'] ) . '"'
 		);
 
-		$error[] = sprintf( _( 'To assign permissions to this user, either change the permissions for this profile using the %s setup or change this user to a user with custom permissions by using %s.' ), ( AllowUse( 'Users/Profiles.php' ) ? '<a href="Modules.php?modname=Users/Profiles.php">' : '' ) . _( 'Profiles' ) . ( AllowUse( 'Users/Profiles.php' ) ? '</a>' : '' ), ( AllowUse( 'Users/User.php' ) ? '<a href="Modules.php?modname=Users/User.php">' : '' ) . _( 'General Info' ) . ( AllowUse( 'Users/User.php' ) ? '</a>' : '' ) );
+		$warning[] = sprintf( _( 'To assign permissions to this user, either change the permissions for this profile using the %s setup or change this user to a user with custom permissions by using %s.' ), ( AllowUse( 'Users/Profiles.php' ) ? '<a href="Modules.php?modname=Users/Profiles.php">' : '' ) . _( 'Profiles' ) . ( AllowUse( 'Users/Profiles.php' ) ? '</a>' : '' ), ( AllowUse( 'Users/User.php' ) ? '<a href="Modules.php?modname=Users/User.php">' : '' ) . _( 'General Info' ) . ( AllowUse( 'Users/User.php' ) ? '</a>' : '' ) );
 
-		echo ErrorMessage( $error );
+		echo ErrorMessage( $warning, 'warning' );
 	}
 }