+++ CHANGES.md
@@ -19,6 +19,7 @@ Changes in 4.3
 - Add SubstitutionsInput & SubstitutionsTextMake functions in Substitutions.fnc.php
 - Use Substitutions functions in CreateParents.php, NotifyParents.php, HonorRoll.php, HonorRoll.fnc.php, Transcripts.php & Letters.php
 - Add referrer to ErrorSendEmail() in ErrorMessage.fnc.php
+- Improve User Permissions program warning & translations in Exceptions.php
 
 Changes in 4.2
 --------------